% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{.start_server}
\alias{.start_server}
\title{Start HTTP server for dashboard}
\usage{
.start_server(port = 8080, docs_dir = "docs")
}
\arguments{
\item{port}{Port for HTTP server}

\item{docs_dir}{Directory with dashboard HTML}
}
\description{
Start HTTP server for dashboard
}
