% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_dbip}
\alias{download_dbip}
\title{Download DB-IP data files}
\usage{
download_dbip()
}
\value{
Character vector with paths to downloaded files
}
\description{
This function downloads geolocation and ASN data from DB-IP website.
}
\keyword{internal}
