% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_test_db.R
\name{make_demo_dashboard}
\alias{make_demo_dashboard}
\title{Create HTML Report with DB-IP Data Visualizations with test data}
\usage{
make_demo_dashboard()
}
\value{
Invisibly returns the path to the created HTML file (`test_docs/test_index.html`)
}
\description{
Renders an interactive HTML report with maps and charts based on DB-IP data.
The report is saved to `docs/index.html` in the current working directory.
}
